# back/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /equipment-types:
    post:
      operationId: EquipmentTypesController_create
      parameters: []
      requestBody:
        required: true
        description: New equipment type payload
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Weapon
              required:
                - name
              additionalProperties: false
      responses:
        '201':
          description: Equipment type created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 7
                  name:
                    type: string
                    example: Weapon
      security:
        - bearer: []
      tags:
        - Equipment Types
    get:
      operationId: EquipmentTypesController_findAll
      parameters: []
      responses:
        '200':
          description: List of equipment types
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 1
                    name:
                      type: string
                      example: Armor
      security:
        - bearer: []
      tags:
        - Equipment Types
  /equipment-types/{id}:
    get:
      operationId: EquipmentTypesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment type ID
          schema:
            example: 7
            type: number
      responses:
        '200':
          description: Equipment type by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 7
                  name:
                    type: string
                    example: Weapon
      security:
        - bearer: []
      tags:
        - Equipment Types
    patch:
      operationId: EquipmentTypesController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment type ID
          schema:
            example: 7
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Ranged Weapon
              additionalProperties: false
      responses:
        '200':
          description: Updated equipment type
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 7
                  name:
                    type: string
                    example: Ranged Weapon
      security:
        - bearer: []
      tags:
        - Equipment Types
    delete:
      operationId: EquipmentTypesController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment type ID
          schema:
            example: 7
            type: number
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 7
                  deleted:
                    type: boolean
                    example: true
      security:
        - bearer: []
      tags:
        - Equipment Types
  /consumables:
    post:
      operationId: ConsumablesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConsumableDto'
      responses:
        '201':
          description: ''
      tags:
        - Consumables
    get:
      operationId: ConsumablesController_findAll
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - Consumables
  /consumables/{id}:
    get:
      operationId: ConsumablesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - Consumables
    patch:
      operationId: ConsumablesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConsumableDto'
      responses:
        '200':
          description: ''
      tags:
        - Consumables
    delete:
      operationId: ConsumablesController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - Consumables
  /consumables/{id}/purchase:
    post:
      operationId: ConsumablesController_purchase
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseConsumableDto'
      responses:
        '201':
          description: ''
      tags:
        - Consumables
...
  /equipment:
    get:
      operationId: EquipmentController_findAll
      parameters: []
      responses:
        '200':
          description: List of equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 10
                    name:
                      type: string
                      example: Longsword
                    equipmentTypeId:
                      type: number
                      example: 7
                    cost:
                      type: number
                      example: 100
                    maxDurability:
                      type: number
                      example: 50
                    currentDurability:
                      type: number
                      example: 50
      security:
        - bearer: []
      tags:
        - Equipment
    post:
      operationId: EquipmentController_create
      parameters: []
      requestBody:
        required: true
        description: New equipment payload
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Longsword
                equipmentTypeId:
                  type: number
                  example: 7
                cost:
                  type: number
                  example: 100
                maxDurability:
                  type: number
                  example: 50
              required:
                - name
                - equipmentTypeId
                - cost
                - maxDurability
              additionalProperties: false
      responses:
        '201':
          description: Equipment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 10
                  name:
                    type: string
                    example: Longsword
                  equipmentTypeId:
                    type: number
                    example: 7
                  cost:
                    type: number
                    example: 100
                  maxDurability:
                    type: number
                    example: 50
                  currentDurability:
                    type: number
                    example: 50
      security:
        - bearer: []
      tags:
        - Equipment
  /equipment/{id}:
    get:
      operationId: EquipmentController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment ID
          schema:
            example: 10
            type: number
      responses:
        '200':
          description: Equipment by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 10
                  name:
                    type: string
                    example: Longsword
                  equipmentTypeId:
                    type: number
                    example: 7
                  cost:
                    type: number
                    example: 100
                  maxDurability:
                    type: number
                    example: 50
                  currentDurability:
                    type: number
                    example: 50
      security:
        - bearer: []
      tags:
        - Equipment
    patch:
      operationId: EquipmentController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment ID
          schema:
            example: 10
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Enchanted Longsword
                equipmentTypeId:
                  type: number
                  example: 8
                cost:
                  type: number
                  example: 150
                maxDurability:
                  type: number
                  example: 60
                currentDurability:
                  type: number
                  example: 45
              additionalProperties: false
      responses:
        '200':
          description: Updated equipment
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 10
                  name:
                    type: string
                    example: Enchanted Longsword
                  equipmentTypeId:
                    type: number
                    example: 8
                  cost:
                    type: number
                    example: 150
                  maxDurability:
                    type: number
                    example: 60
                  currentDurability:
                    type: number
                    example: 45
      security:
        - bearer: []
      tags:
        - Equipment
    delete:
      operationId: EquipmentController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment ID
          schema:
            example: 10
            type: number
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 10
                  deleted:
                    type: boolean
                    example: true
      security:
        - bearer: []
      tags:
        - Equipment
...
info:
  title: API "Les kips AH"
  description: |-
    This is the API information regarding the project "Les capuches d'Opale".

    C'est un outil de gestion d'aventurier pour une guilde.
  version: 1.0.0
  contact: {}
  termsOfService: https://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http